+++ src/DDTrace/Integrations/Lumen/V5/LumenIntegrationLoader.php
@@ -109,9 +109,6 @@ final class LumenIntegrationLoader
      */
     private function shouldLoad()
     {
-        if ('cli' === PHP_SAPI && 'dd_testing' !== getenv('APP_ENV')) {
-            return false;
-        }
         if (!Configuration::get()->isIntegrationEnabled(LumenIntegration::NAME)) {
             return false;
         }